Safari-611.1.10/LayoutTests/css3/calc/calc-errors-expected.txt

All boxes below should be 100px * 100px and green.

unclosed calc => PASS
unclosed calc with garbage => PASS
garbage => PASS
zero division => PASS
non length => PASS
number + length => PASS
length + number => PASS
percent + number => PASS
number + percent => PASS
length - number => PASS
number - length => PASS
percent - number => PASS
number - percent => PASS
length * length => PASS
length * percent => PASS
percent * length => PASS
percent * percent => PASS
number / length => PASS
number / percent => PASS
length / length => PASS
length / percent => PASS
percent / length => PASS
percent / percent => PASS
number mod length => PASS
number mod percent => PASS
length mod length => PASS
length mod percent => PASS
percent mod length => PASS
percent mod percent => PASS
mod10 => PASS
1mod => PASS
70px+40px no whitespace around + => PASS
70px +40px no whitespace on right of + => PASS
70px+ 40px no whitespace on left of + => PASS
70px+-40px no whitespace around + => PASS
70px-40px no whitespace around - => PASS
70px -40px no whitespace on right of - => PASS
70px- 40px no whitespace on left of - => PASS
70px-+40px no whitespace around - => PASS
too many nests => PASS
end with operator => PASS
start with operator => PASS
no expressions => PASS
too many pluses => PASS
no binary operator => PASS
two binary operators => PASS
invalid operator '@' => PASS
invalid operator 'flim' => PASS
invalid operator '@' => PASS
invalid operator 'flim' => PASS
invalid operator 'flim' with parens => PASS
